requirements.txt
streamlit
transformers
torch
edge-tts
easyocr
langdetect
streamlit-folium
streamlit-mic-recorder
pillow
numpy
python-docx
pymupdf